# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tile Map API",
        "description": "API for managing tile maps and their layers",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Base URL"
        }
    ],
    "paths": {
...
        "/tile-sets": {
            "get": {
                "tags": [
                    "Tile Sets"
                ],
                "summary": "Get all tile sets",
                "description": "Get all tile sets",
                "operationId": "bb944de821d7b8e00c9685276209ced4",
                "responses": {
                    "200": {
                        "description": "List of tile sets",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TileSet"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tile Sets"
                ],
                "summary": "Create a new tile set",
                "description": "Create a new tile set",
                "operationId": "429350cb4f811d952b289a6bdbc411f1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TileSetStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tile set created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TileSet"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tile-sets/{uuid}": {
            "get": {
                "tags": [
                    "Tile Sets"
                ],
                "summary": "Get a specific tile set",
                "description": "Get a specific tile set",
                "operationId": "2ced28a7f9a019b62abcfc37f95dda82",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the tile set",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tile set details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TileSet"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tile set not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Tile Sets"
                ],
                "summary": "Update a tile set",
                "description": "Update a tile set",
                "operationId": "005caffc9994a81b49553226d8440c57",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the tile set",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TileSetUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tile set updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TileSet"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tile set not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "TileSet": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "uuid": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "name": {
                        "type": "string",
                        "example": "My Tile Set"
                    },
                    "image_width": {
                        "type": "integer",
                        "example": 512
                    },
                    "image_height": {
                        "type": "integer",
                        "example": 512
                    },
                    "tile_width": {
                        "type": "integer",
                        "example": 32
                    },
                    "tile_height": {
                        "type": "integer",
                        "example": 32
                    },
                    "tile_count": {
                        "type": "integer",
                        "example": 256
                    },
                    "first_gid": {
                        "type": "integer",
                        "example": 1
                    },
                    "margin": {
                        "type": "integer",
                        "example": 0
                    },
                    "spacing": {
                        "type": "integer",
                        "example": 0
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "TileSetStoreRequest": {
                "required": [
                    "name",
                    "imageWidth",
                    "imageHeight",
                    "tileWidth",
                    "tileHeight",
                    "tileCount",
                    "firstGid"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "My New Tile Set"
                    },
                    "imageWidth": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 512
                    },
                    "imageHeight": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 512
                    },
                    "tileWidth": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 32
                    },
                    "tileHeight": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 32
                    },
                    "tileCount": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 256
                    },
                    "firstGid": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 1
                    },
                    "margin": {
                        "type": "integer",
                        "minimum": 0,
                        "example": 0
                    },
                    "spacing": {
                        "type": "integer",
                        "minimum": 0,
                        "example": 0
                    }
                },
                "type": "object"
            },
            "TileSetUpdateRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Updated Tile Set Name"
                    },
                    "imageWidth": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 1024
                    },
                    "imageHeight": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 1024
                    },
                    "tileWidth": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 64
                    },
                    "tileHeight": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 64
                    },
                    "tileCount": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 256
                    },
                    "firstGid": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 1
                    },
                    "margin": {
                        "type": "integer",
                        "minimum": 0,
                        "example": 2
                    },
                    "spacing": {
                        "type": "integer",
                        "minimum": 0,
                        "example": 2
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}